+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.7.0] - 2020-05-27
+- added open url option to push notification click event
+- added dca picker for news to open url on click field
+- fixed error row in send command
+
 ## [0.6.2] - 2019-11-14
 * added service definitions for symfony 4